t/data/parsing/temma-inclusion-3.dat

#data
<t:content><link>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <temma content>
|     <link>

#data
<p><t:content><link>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <p>
|       <temma content>
|       <link>
